requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
streamlit>=1.28.0
PyPDF2>=3.0.0
torch>=2.0.0
